# Conflicts:
#	include/Engine/Network/Client.h
#	include/Engine/Network/Network.h
#	include/Engine/Network/Packet.h
#	include/Engine/Network/Server.h
#	src/Engine/Network/Client.cpp
#	src/Engine/Network/Packet.cpp
#	src/Engine/Network/Server.cpp
#	src/Tests/HealthSystemTest.cpp

+++ include/Engine/Network/Network.h
@@ -29,7 +29,7 @@ protected:
     unsigned int m_SaveDataIntervalMs = 1000;
     std::clock_t m_SaveDataTimer;
     unsigned int m_MaxConnections;
-    unsigned int m_TimeoutMs;
+    double m_TimeoutMs;
     void logSentData(int bytesSent);
     void logReceivedData(int bytesReceived);
     void saveToFile();
