+++ include/Engine/Network/Server.h
@@ -35,15 +35,15 @@ protected:
     std::map<PlayerID, PlayerDefinition> m_ConnectedPlayers;
     // HACK: Fix INPUTSIZE
     char readBuffer[BUFFERSIZE] = { 0 };
-    int bytesRead = 0;
+    size_t bytesRead = 0;
     // time for previouse message
     std::clock_t previousePingMessage = std::clock();
     std::clock_t previousSnapshotMessage = std::clock();
     std::clock_t timOutTimer = std::clock();
 
     // How often we send messages (milliseconds)
-    int pingIntervalMs;
-    int snapshotInterval;
+    float pingIntervalMs;
+    float snapshotInterval;
     int checkTimeOutInterval = 100;
     int m_NextPlayerID = 0;
     //Timers
